requirements.txt
Flask==2.3.2
requests==2.31.0
Pillow==10.0.1
python-dotenv==1.0.0
dashscope==1.13.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
numpy>=1.24.0